# frontend/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Harness API",
    "version": "0.3.0"
  },
  "paths": {
    "/api/v1/runs": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Runs",
        "description": "List all runs with folded state summary, ordered by most recent.",
        "operationId": "list_runs_api_v1_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Create Run",
        "description": "Create a new run and write the RunStarted event.",
        "operationId": "create_run_api_v1_runs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run",
        "description": "Get the folded state snapshot of a single run.",
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "runs"
        ],
        "summary": "Delete Run",
        "description": "Cancel a run and write a RunFailed termination event.",
        "operationId": "delete_run_api_v1_runs__run_id__delete",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/events": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run Events",
        "description": "Get the event stream for a run, optionally from a given seq.",
        "operationId": "get_run_events_api_v1_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "from_seq",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "From Seq"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/pause": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Pause Run",
        "description": "Pause a running run. Writes RunPaused event and halts the scheduler loop.",
        "operationId": "pause_run_api_v1_runs__run_id__pause_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/PauseRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/resume": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Resume Run",
        "description": "Resume a paused run. Requires an active scheduler instance.",
        "operationId": "resume_run_api_v1_runs__run_id__resume_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/confirm": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Confirm Run",
        "description": "Submit an operator confirmation decision. Idempotent per confirmation_id.",
        "operationId": "confirm_run_api_v1_runs__run_id__confirm_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ConfirmRequest": {
        "properties": {
          "confirmation_id": {
            "type": "string",
            "title": "Confirmation Id"
          },
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed"
          },
          "operator_id": {
            "type": "string",
            "title": "Operator Id",
            "default": "operator"
          }
        },
        "type": "object",
        "required": [
          "confirmation_id",
          "confirmed"
        ],
        "title": "ConfirmRequest"
      },
      "CreateRunRequest": {
        "properties": {
          "intent": {
            "type": "string",
            "title": "Intent"
          }
        },
        "type": "object",
        "required": [
          "intent"
        ],
        "title": "CreateRunRequest"
      },
      "EventListResponse": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventResponse"
            },
            "type": "array",
            "title": "Events"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "events",
          "total"
        ],
        "title": "EventListResponse"
      },
      "EventResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "seq",
          "event_type",
          "payload",
          "created_at"
        ],
        "title": "EventResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PauseRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": "user_requested"
          }
        },
        "type": "object",
        "title": "PauseRequest"
      },
      "RunDetailResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "intent": {
            "type": "string",
            "title": "Intent"
          },
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "event_count": {
            "type": "integer",
            "title": "Event Count"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "pause_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pause Reason"
          },
          "pending_confirmations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Pending Confirmations",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status",
          "intent",
          "seq",
          "event_count"
        ],
        "title": "RunDetailResponse"
      },
      "RunListResponse": {
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "type": "array",
            "title": "Runs"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "runs",
          "total"
        ],
        "title": "RunListResponse"
      },
      "RunSummary": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "intent": {
            "type": "string",
            "title": "Intent",
            "default": ""
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "running"
          },
          "event_count": {
            "type": "integer",
            "title": "Event Count",
            "default": 0
          },
          "created_at": {
            "type": "number",
            "title": "Created At",
            "default": 0
          },
          "updated_at": {
            "type": "number",
            "title": "Updated At",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "run_id"
        ],
        "title": "RunSummary"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}